fix region value
